# Optimizations to Schlinkert algorithm, up to -87.5% reduction