- No clone/to_owned
- Combined loops
- No repeated calls to `generate_cn` with the same `n`. Removed duplicate work. This was the largest single benefit.

cargo bench
===
```
     Running benches/long_not_ud_list.rs (target/release/deps/long_not_ud_list-f027967d24d31a17)
Gnuplot not found, using plotters backend
Benchmarking Check long, not UD list/Schlinkert: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 47.2s.
Check long, not UD list/Schlinkert
                        time:   [4.6974 s 4.7713 s 4.8589 s]
                        change: [-83.954% -83.690% -83.388%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high mild

     Running benches/long_ud_list.rs (target/release/deps/long_ud_list-3022173e81445c0e)
Gnuplot not found, using plotters backend
Benchmarking Check long, UD list/Schlinkert: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 71.2s.
Check long, UD list/Schlinkert
                        time:   [7.0938 s 7.1845 s 7.3075 s]
                        change: [-87.711% -87.545% -87.338%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high severe

     Running benches/medium_list.rs (target/release/deps/medium_list-59159e0384472fee)
Gnuplot not found, using plotters backend
Benchmarking Check medium, not UD list/Schlinkert: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 8.2s or enable flat sampling.
Check medium, not UD list/Schlinkert
                        time:   [148.75 ms 155.88 ms 164.69 ms]
                        change: [-75.701% -74.607% -73.357%] (p = 0.00 < 0.10)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) high mild

     Running benches/short_list.rs (target/release/deps/short_list-51c6e1773d8596ab)
Gnuplot not found, using plotters backend
Check short UD list/Schlinkert
                        time:   [977.15 µs 979.91 µs 982.85 µs]
                        change: [-24.010% -23.716% -23.425%] (p = 0.00 < 0.10)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe